refactor(folios): refactor index route workflow
