+++ app/controllers/product/main_folios_controller.ts
@@ -1,0 +1,4 @@
+import { inject } from '@adonisjs/core'
+
+@inject()
+export default class MainFoliosController {}
