+++ start/routes.ts
@@ -6,6 +6,7 @@ const ApiInfoController = () => import('#controllers/system/api_info_controller'
 const CardsController = () => import('#controllers/product/cards_controller')
 const ScanController = () => import('#controllers/product/scan_controller')
 const FoliosController = () => import('#controllers/product/folios_controller')
+const ChildFoliosController = () => import('#controllers/product/child_folios_controller')
 const CardFoliosController = () => import('#controllers/product/card_folios_controller')
 const SetsController = () => import('#controllers/product/sets_controller')
 const FiltersController = () => import('#controllers/product/filters_controller')
@@ -28,7 +29,7 @@ router
           .prefix('cards')
         router
           .group(() => {
-            router.get('/', [FoliosController, 'index']).use(middleware.auth())
+            router.get('/', [ChildFoliosController, 'index']).use(middleware.auth())
             router.get('/cards', [FoliosController, 'mainFolioCards']).use(middleware.auth())
             router.get('/statistics', [FoliosController, 'statistics']).use(middleware.auth())
             router.get('/:id', [FoliosController, 'show']).use(middleware.auth())
