+++ tests/functional/controllers/main_folios_controller.spec.ts
@@ -1,0 +1,18 @@
+import { test } from '@japa/runner'
+import testUtils from '@adonisjs/core/services/test_utils'
+import sinon from 'sinon'
+import AuthServiceMock from '#tests/mocks/auth_service_mock'
+
+test.group('Main folio controller', (group) => {
+  let wardenApiClientStub: sinon.SinonStub
+
+  group.setup(() => {
+    wardenApiClientStub = AuthServiceMock.setupWardenApiClientStub()
+  })
+
+  group.each.setup(() => testUtils.db().withGlobalTransaction())
+
+  group.teardown(() => {
+    wardenApiClientStub.restore()
+  })
+})
